Consolidate PDP metrics into remoteblockstore package
- Move PDP metrics from pdp/ to market/retrieval/remoteblockstore/
- Update piecehandler to use consolidated metrics functions
- Implement isPDPPiece method directly in Provider
- Remove redundant pdp/metrics.go file
- Maintain same metric names and functionality for monitoring
